+++ gluecodium/src/test/resources/smoke/comments/output/dart/lib/src/smoke/comments_links.dart
@@ -37,6 +37,7 @@ abstract class CommentsLinks {
   /// * top level enum: [TypeCollectionEnum]
   /// * top level enum item: [TypeCollectionEnum.item]
   /// * error: [Comments_SomethingWrongException]
+  /// * lambda: [Comments_SomeLambda]
   /// * type from aux sources, same package: [AuxClass]
   /// * type from aux sources, different package: [AuxStruct]
   ///   * we can also have