+++ gluecodium/src/test/resources/smoke/comments/output/swift/smoke/CommentsLinks.swift
@@ -53,6 +53,7 @@ public class CommentsLinks {
     /// * top level enum: `TypeCollectionEnum`
     /// * top level enum item: `TypeCollectionEnum.item`
     /// * error: `Comments.SomethingWrongError`
+    /// * lambda: `Comments.SomeLambda`
     /// * type from aux sources, same package: `AuxClass`
     /// * type from aux sources, different package: `AuxStruct`
     ///     * we can also have